+++ gson2/src/main/java/org/jdbi/v3/gson2/Gson2Plugin.java
@@ -31,6 +31,6 @@ public class Gson2Plugin implements JdbiPlugin {
     @Override
     public void customizeJdbi(Jdbi jdbi) {
         jdbi.installPlugin(new JsonPlugin());
-        jdbi.getConfig(JsonConfig.class).setJsonMapper(new GsonJsonImpl());
+        jdbi.getConfig(JsonConfig.class).setJsonMapper(new GsonJsonMapper());
     }
 }
