+++ jackson2/src/main/java/org/jdbi/v3/jackson2/Jackson2Plugin.java
@@ -31,6 +31,6 @@ public class Jackson2Plugin implements JdbiPlugin {
     @Override
     public void customizeJdbi(Jdbi jdbi) {
         jdbi.installPlugin(new JsonPlugin());
-        jdbi.getConfig(JsonConfig.class).setJsonMapper(new JacksonJsonImpl());
+        jdbi.getConfig(JsonConfig.class).setJsonMapper(new JacksonJsonMapper());
     }
 }
